Event: Make `description`, `local_date`, `local_time` optional
After making a request to the API and getting a real JSON response
again, it turns out that the above mentioned fields are actually
optional, and will break Serde's parsing when they're not in the JSON
string coming from Meetup. Make these fields `Option`s so the program
doesn't break when they're not there.